Updated register and login route
